+++ test/integration/03-integration-peers.js
@@ -8,10 +8,10 @@ var libDir = require('../_lib/lib-dir');
 var baseConfig = require('../_lib/base-config');
 var stopCluster = require('../_lib/stop-cluster');
 
-describe('03 - integration - peers', function () {
+describe(require('../_lib/test-helper').testName(__filename, 3), function () {
 
   this.timeout(20000);
-  
+
   var servers, localInstance;
 
   function localInstanceConfig(seq) {