+++ test/unit/02-broker-component.js
@@ -9,8 +9,16 @@ describe('02 - unit - brokerage component', function() {
   it('injects and detaches the brokerage component', function(done) {
     //package, mesh, client
     var mockModels = {};
-    var mockMesh = {};
-    var mockClient = {};
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+    var mockClient = {
+      on:function(){}
+    };
 
     var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient);
 
@@ -23,7 +31,7 @@ describe('02 - unit - brokerage component', function() {
     });
   });
 
-  it('tests the __checkDuplicateInjections', function(done) {
+  it('tests the __checkDuplicateInjections method', function(done) {
 
     //package, mesh, client
     var mockModels = {
@@ -39,13 +47,243 @@ describe('02 - unit - brokerage component', function() {
       }
     };
 
-    var mockMesh = {};
-    var mockClient = {};
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
     var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient);
 
     brokerage.inject(function(e){
       expect(e.toString()).to.be('Error: Duplicate attempts to broker the remoteComponent3 component by brokerComponent & brokerComponent1');
       done();
+    });
+  });
+
+  it('tests the instance method', function() {
+
+    //package, mesh, client
+    var mockModels = {
+      brokerComponent: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      },
+      brokerComponent1: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      }
+    };
+
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
+
+    var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient);
+    expect(require('../../lib/brokerage').instance('mock')).to.eql(brokerage);
+  });
+
+  it('tests the deferProxyStart method', function(done) {
+
+    //package, mesh, client
+    var mockModels = {
+      brokerComponent: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      },
+      brokerComponent1: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      }
+    };
+
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
+    var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient);
+    brokerage.deferProxyStart({test:'proxy'}).then(function(){
+      expect(brokerage.__proxy).to.eql({test:'proxy'});
+      done();
+    });
+  });
+
+  it('tests the dependenciesSatisfied method', function() {
+
+    //package, mesh, client
+    var mockModels = {
+      brokerComponent: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      },
+      brokerComponent1: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      }
+    };
+
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
+
+    var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient);
+    brokerage.__satisfiedElementNames = ['test1', 'test2'];
+    brokerage.__injectedElementNames = ['test2', 'test1'];
+    expect(brokerage.dependenciesSatisfied()).to.be(true);
+
+    brokerage.__satisfiedElementNames = ['test1', 'test2'];
+    brokerage.__injectedElementNames = ['test2'];
+    expect(brokerage.dependenciesSatisfied()).to.be(false);
+  });
+
+  it('tests the __checkDependenciesSatisfied method', function(done) {
+
+    //package, mesh, client
+    var mockModels = {
+      brokerComponent: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      },
+      brokerComponent1: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      }
+    };
+
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
+
+    var mockLogger = {
+
+    };
+
+    var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient, mockLogger, {
+      dependenciesSatisfiedDeferListen:true
+    });
+    brokerage.__proxy = {
+      start:done
+    };
+    brokerage.__satisfiedElementNames = ['test2'];
+    brokerage.__injectedElementNames = ['test2', 'test1'];
+    brokerage.__checkDependenciesSatisfied();
+
+    brokerage.__satisfiedElementNames = ['test1', 'test2'];
+    brokerage.__injectedElementNames = ['test2', 'test1'];
+    brokerage.__checkDependenciesSatisfied();
+  });
+
+  it('tests the __handleDependencyMet method', function(done) {
+
+    //package, mesh, client
+    var mockModels = {
+      brokerComponent: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      },
+      brokerComponent1: {
+        remoteComponent3: {
+          version: '^2.0.0'
+        }
+      }
+    };
+
+    var mockMesh = {
+      _mesh:{
+        config:{
+          name:'mock'
+        }
+      }
+    };
+
+    var mockClient = {
+      on:function(){}
+    };
+
+    var mockLogger = {
+      info:function(msg){
+        expect('element re-injected: test').to.be(msg);
+        done();
+      },
+      error:function(){
+        done(arguments[0]);
+      }
+    };
+
+    var brokerage = require('../../lib/brokerage').create(mockModels, mockMesh, mockClient, mockLogger);
+
+    brokerage.__client = {
+      construct:function(){
+        return {
+          exchange:{}
+        };
+      }
+    };
+
+    brokerage.__injectedElements = [
+      {
+        component:{
+          name: 'test'
+        }
+      }
+    ];
+
+    brokerage.__mesh = {
+      _updateElement:function(){
+        return new Promise((resolve)=>{
+          resolve();
+        });
+      }
+    };
+
+    brokerage.__handleDependencyMet({
+      componentName:'test',
+      description:{}
     });
   });
 });